chore(deps): enable automatic updates for GitHub Actions via Dependabot (#187)
### 💚 Dependabot automatic updates
This PR enables automatic updates for GitHub Actions via
[Dependabot](https://docs.github.com/en/code-security/dependabot/working-with-dependabot/keeping-your-actions-up-to-date-with-dependabot).

Dependabot will periodically check for new versions of the actions and
create a PR to update the version used in the repository.

This should help keeping your GitHub Actions up to date with the latest
versions of the actions.

#### 🔍 How was this detected and generated?
This PR was generated using
[ethpandaops/github-actions-checker](https://github.com/ethpandaops/github-actions-checker).